Mobile Slider Menu
Added the html and styling for the left slideing burger menu, as well as a background scrim possibility for the background when the menu is open.
still need to create a function for being able to drag the menu onto the screen, if that's even possible on a browser, and how to use the scrim (eg. skrim blocks out rest of body, even if opacity is 0, so no clickable buttons etc. but there is not a possibility for transition animations of a display attribute. Issues issues )